Fix MSVC compiler warnings
reg. conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/layout/v0/beamlayout.cpp
@@ -992,7 +992,7 @@ void BeamLayout::createBeamSegments(Beam* item, LayoutContext& ctx, const std::v
                     if (startCr == endCr && startCr->isChord()) {
                         bool isBeamletBefore = calcIsBeamletBefore(item,
                                                                    toChord(startCr),
-                                                                   beamletIndex,
+                                                                   static_cast<int>(beamletIndex),
                                                                    level,
                                                                    previousBreak16,
                                                                    previousBreak32);
